Fixed movement bug
Because of the delay at the end of the loop, the program missed key presses and the snake didn't behave as expected. Removing the delay function and adding a timer to control when to move the snake and cycle through the loop faster to be able to detect keystrokes fixed the problem.